+++ meta_subject/action_items.py
@@ -1,6 +1,7 @@
 from edc_adverse_event.constants import AE_INITIAL_ACTION
 from edc_action_item import Action, site_action_items
-from edc_constants.constants import YES, HIGH_PRIORITY
+from edc_constants.constants import NONE, YES, HIGH_PRIORITY
+from edc_reportable import GRADE3, GRADE4
 from meta_visit_schedule.constants import DAY1
 
 from .constants import (
@@ -11,7 +12,31 @@ from .constants import (
     BLOOD_RESULTS_RFT_ACTION,
     BLOOD_RESULTS_HBA1C_ACTION,
     BLOOD_RESULTS_EGFR_ACTION,
+    FOLLOWUP_EXAMINATION_ACTION,
 )
+
+
+def is_baseline(action):
+    return action.reference_obj.subject_visit.appointment.visit_code == DAY1
+
+
+class FollowupExaminationAction(Action):
+    name = FOLLOWUP_EXAMINATION_ACTION
+    priority = HIGH_PRIORITY
+    display_name = "Followup Exam: AE"
+    reference_model = "meta_subject.followupexamination"
+    show_on_dashboard = True
+    create_by_user = False
+
+    def get_next_actions(self):
+        next_actions = []
+        if (
+            self.reference_obj.symptoms_g3.exclude(name=NONE).count() > 0
+            or self.reference_obj.symptoms_g4.exclude(name=NONE).count() > 0
+            or self.reference_obj.any_other_problems_sae_grade in [GRADE3, GRADE4]
+        ) and not is_baseline(self):
+            next_actions.append(AE_INITIAL_ACTION)
+        return next_actions
 
 
 class BaseBloodResultsAction(Action):
@@ -26,16 +51,12 @@ class BaseBloodResultsAction(Action):
     def reopen_action_item_on_change(self):
         return False
 
-    @property
-    def is_baseline(self):
-        return self.reference_obj.subject_visit.appointment.visit_code == DAY1
-
     def get_next_actions(self):
         next_actions = []
         if (
             self.reference_obj.results_abnormal == YES
             and self.reference_obj.results_reportable == YES
-            and not self.is_baseline
+            and not is_baseline(self)
         ):
             # AE for reportable result, though not on DAY1.0
             next_actions = [AE_INITIAL_ACTION]
@@ -67,7 +88,7 @@ class BloodResultsLipidAction(BaseBloodResultsAction):
 
 
 class BloodResultsEgfrAction(BaseBloodResultsAction):
-    name = (BLOOD_RESULTS_EGFR_ACTION,)
+    name = BLOOD_RESULTS_EGFR_ACTION
     display_name = "Reportable eGFR"
     reference_model = "meta_subject.bloodresultsfbc"
 
@@ -91,3 +112,4 @@ site_action_items.register(BloodResultsLftAction)
 site_action_items.register(BloodResultsRftAction)
 site_action_items.register(BloodResultsGluAction)
 site_action_items.register(BloodResultsHba1cAction)
+site_action_items.register(FollowupExaminationAction)