+++ meta_subject/admin/__init__.py
@@ -1,3 +1,4 @@
+from .autocomplete_admin import ArvRegimensAdmin
 from .blood_results import (
     BloodResultsGluAdmin,
     BloodResultsFbcAdmin,
@@ -5,15 +6,15 @@ from .blood_results import (
     BloodResultsLftAdmin,
 )
 from .complications_admin import ComplicationsAdmin
-from .followup_admin import FollowupAdmin
+from .followup_examination_admin import FollowupExaminationAdmin
 from .followup_vitals_admin import FollowupVitalsAdmin
 from .glucose_admin import GlucoseAdmin
 from .health_economics_admin import HealthEconomicsAdmin
 from .malaria_test_admin import MalariaTestAdmin
 from .medication_adherence_admin import MedicationAdherenceAdmin
-from .missed_visit_admin import MissedVisitAdmin
 from .patient_history_admin import PatientHistoryAdmin
 from .physical_exam_admin import PhysicalExamAdmin
 from .subject_requisition_admin import SubjectRequisitionAdmin
 from .subject_visit_admin import SubjectVisitAdmin
+from .subject_visit_missed_admin import SubjectVisitMissedAdmin
 from .urine_dipstick_test_admin import UrineDipstickTestAdmin