+++ meta_subject/forms/__init__.py
@@ -6,14 +6,14 @@ from .blood_results import (
     BloodResultsFbcForm,
     BloodResultsHba1cForm,
 )
-from .followup_form import FollowupForm
+from .followup_examination_form import FollowupExaminationForm
 from .followup_vitals_form import FollowupVitalsForm
 from .glucose_form import GlucoseForm
 from .malaria_test_form import MalariaTestForm
 from .medication_adherence_form import MedicationAdherenceForm
-from .missed_visit_form import MissedVisitForm
 from .patient_history_form import PatientHistoryForm
 from .physical_exam_form import PhysicalExamForm
 from .subject_requisition_form import SubjectRequisitionForm
 from .subject_visit_form import SubjectVisitForm
 from .urine_dipstick_test_form import UrineDipstickTestForm
+from .subject_visit_missed_form import SubjectVisitMissedForm