+++ meta_subject/models/__init__.py
@@ -7,15 +7,15 @@ from .blood_results import (
     BloodResultsGlu,
 )
 from .complications import Complications
-from .followup import Followup
+from .followup_examination import FollowupExamination
 from .followup_vitals import FollowupVitals
 from .glucose import Glucose
 from .health_economics import HealthEconomics
 from .malaria_test import MalariaTest
 from .medication_adherence import MedicationAdherence
-from .missed_visit import MissedVisit
 from .patient_history import PatientHistory
 from .physical_exam import PhysicalExam
+from .subject_visit_missed import SubjectVisitMissed
 from .subject_requisition import SubjectRequisition
 from .subject_visit import SubjectVisit
 from .urine_dipstick_test import UrineDipstickTest