+++ meta_subject/models/blood_results/blood_results_fbc.py
@@ -3,7 +3,7 @@ from django.db import models
 from django.db.models.deletion import PROTECT
 from edc_constants.choices import YES_NO
 from edc_crf.model_mixins import CrfNoManagerModelMixin
-from edc_model.models import BaseUuidModel
+from edc_model import models as edc_models
 from edc_model.validators import datetime_not_future
 from edc_reportable import (
     GRAMS_PER_DECILITER,
@@ -19,7 +19,9 @@ from ...constants import BLOOD_RESULTS_FBC_ACTION
 from ..subject_requisition import SubjectRequisition
 
 
-class BloodResultsFbc(CrfNoManagerModelMixin, BloodResultsModelMixin, BaseUuidModel):
+class BloodResultsFbc(
+    CrfNoManagerModelMixin, BloodResultsModelMixin, edc_models.BaseUuidModel
+):
 
     action_name = BLOOD_RESULTS_FBC_ACTION
 
@@ -189,6 +191,6 @@ class BloodResultsFbc(CrfNoManagerModelMixin, BloodResultsModelMixin, BaseUuidMo
         blank=True,
     )
 
-    class Meta(CrfNoManagerModelMixin.Meta):
+    class Meta(CrfNoManagerModelMixin.Meta, edc_models.BaseUuidModel.Meta):
         verbose_name = "Blood Result: FBC"
         verbose_name_plural = "Blood Results: FBC"