+++ meta_subject/models/blood_results/blood_results_lipid.py
@@ -3,7 +3,7 @@ from django.db import models
 from django.db.models.deletion import PROTECT
 from edc_constants.choices import YES_NO
 from edc_crf.model_mixins import CrfNoManagerModelMixin
-from edc_model.models import BaseUuidModel
+from edc_model import models as edc_models
 from edc_model.validators import datetime_not_future
 from edc_reportable.choices import REPORTABLE
 from edc_reportable.units import MILLIMOLES_PER_LITER
@@ -13,7 +13,9 @@ from ...constants import BLOOD_RESULTS_LIPID_ACTION
 from ..subject_requisition import SubjectRequisition
 
 
-class BloodResultsLipid(CrfNoManagerModelMixin, BloodResultsModelMixin, BaseUuidModel):
+class BloodResultsLipid(
+    CrfNoManagerModelMixin, BloodResultsModelMixin, edc_models.BaseUuidModel
+):
 
     action_name = BLOOD_RESULTS_LIPID_ACTION
 
@@ -126,6 +128,6 @@ class BloodResultsLipid(CrfNoManagerModelMixin, BloodResultsModelMixin, BaseUuid
         blank=True,
     )
 
-    class Meta(CrfNoManagerModelMixin.Meta):
+    class Meta(CrfNoManagerModelMixin.Meta, edc_models.BaseUuidModel.Meta):
         verbose_name = "Blood Result: Lipids"
         verbose_name_plural = "Blood Results: Lipids"