+++ meta_subject/models/urine_dipstick_test.py
@@ -2,10 +2,10 @@ from django.db import models
 from edc_crf.model_mixins import CrfModelMixin
 from edc_constants.choices import YES_NO, PRESENT_ABSENT_NA
 from edc_constants.constants import NOT_APPLICABLE
-from edc_model.models.base_uuid_model import BaseUuidModel
+from edc_model import models as edc_models
 
 
-class UrineDipstickTest(CrfModelMixin, BaseUuidModel):
+class UrineDipstickTest(CrfModelMixin, edc_models.BaseUuidModel):
 
     performed = models.CharField(
         verbose_name="Was the urine dipstick test performed?",
@@ -38,5 +38,5 @@ class UrineDipstickTest(CrfModelMixin, BaseUuidModel):
         default=NOT_APPLICABLE,
     )
 
-    class Meta(CrfModelMixin.Meta):
+    class Meta(CrfModelMixin.Meta, edc_models.BaseUuidModel.Meta):
         verbose_name = "Urine Dipstick Test"