+++ meta_visit_schedule/visit_schedules/crfs.py
@@ -15,13 +15,13 @@ crfs_prn = FormsCollection(
 
 crfs_unscheduled = FormsCollection(
     Crf(show_order=10, model="meta_subject.followupvitals"),
-    Crf(show_order=20, model="meta_subject.followup"),
+    Crf(show_order=20, model="meta_subject.followupexamination"),
     Crf(show_order=30, model="meta_subject.medicationadherence"),
     name="unscheduled",
 )
 
 crfs_missed = FormsCollection(
-    Crf(show_order=10, model="meta_subject.missedvisit"), name="missed",
+    Crf(show_order=10, model="meta_subject.subjectvisitmissed"), name="missed",
 )
 
 crfs_d1 = FormsCollection(
@@ -38,7 +38,7 @@ crfs_d1 = FormsCollection(
 
 crfs_w2 = FormsCollection(
     Crf(show_order=10, model="meta_subject.followupvitals"),
-    Crf(show_order=15, model="meta_subject.followup"),
+    Crf(show_order=15, model="meta_subject.followupexamination"),
     Crf(show_order=20, model="meta_subject.healtheconomics"),
     Crf(show_order=30, model="meta_subject.medicationadherence"),
     name="week2",
@@ -46,14 +46,14 @@ crfs_w2 = FormsCollection(
 
 crfs_1m = FormsCollection(
     Crf(show_order=10, model="meta_subject.followupvitals"),
-    Crf(show_order=15, model="meta_subject.followup"),
+    Crf(show_order=15, model="meta_subject.followupexamination"),
     Crf(show_order=20, model="meta_subject.medicationadherence"),
     name="1m",
 )
 
 crfs_3m = FormsCollection(
     Crf(show_order=10, model="meta_subject.followupvitals"),
-    Crf(show_order=15, model="meta_subject.followup"),
+    Crf(show_order=15, model="meta_subject.followupexamination"),
     Crf(show_order=20, model="meta_subject.medicationadherence"),
     Crf(show_order=30, model="meta_subject.bloodresultslft"),
     Crf(show_order=40, model="meta_subject.bloodresultsrft"),
@@ -62,7 +62,7 @@ crfs_3m = FormsCollection(
 
 crfs_6m = FormsCollection(
     Crf(show_order=10, model="meta_subject.followupvitals"),
-    Crf(show_order=15, model="meta_subject.followup"),
+    Crf(show_order=15, model="meta_subject.followupexamination"),
     Crf(show_order=20, model="meta_subject.medicationadherence"),
     Crf(show_order=25, model="meta_subject.glucose"),
     Crf(show_order=40, model="meta_subject.bloodresultshba1c"),
@@ -73,7 +73,7 @@ crfs_6m = FormsCollection(
 
 crfs_9m = FormsCollection(
     Crf(show_order=10, model="meta_subject.followupvitals"),
-    Crf(show_order=15, model="meta_subject.followup"),
+    Crf(show_order=15, model="meta_subject.followupexamination"),
     Crf(show_order=20, model="meta_subject.medicationadherence"),
     Crf(show_order=30, model="meta_subject.bloodresultslft"),
     Crf(show_order=40, model="meta_subject.bloodresultsrft"),
@@ -82,7 +82,7 @@ crfs_9m = FormsCollection(
 
 crfs_12m = FormsCollection(
     Crf(show_order=10, model="meta_subject.followupvitals"),
-    Crf(show_order=15, model="meta_subject.followup"),
+    Crf(show_order=15, model="meta_subject.followupexamination"),
     Crf(show_order=20, model="meta_subject.medicationadherence"),
     Crf(show_order=30, model="meta_subject.glucose"),
     Crf(show_order=40, model="meta_subject.bloodresultshba1c"),